+++ src/pygeoprocessing/geoprocessing.py
@@ -484,7 +484,7 @@ def align_and_resize_raster_stack(
         target_pixel_size, bounding_box_mode, base_vector_path_list=None,
         raster_align_index=None, base_sr_wkt_list=None, target_sr_wkt=None,
         gtiff_creation_options=DEFAULT_GTIFF_CREATION_OPTIONS,
-        vector_mask_options=None):
+        vector_mask_options=None, gdal_warp_options=None):
     """Generate rasters from a base such that they align geospatially.
 
     This function resizes base rasters that are in the same geospatial
@@ -767,7 +767,8 @@ def align_and_resize_raster_stack(
                     'base_sr_wkt': (
                         None if not base_sr_wkt_list else
                         base_sr_wkt_list[index]),
-                    'vector_mask_options': vector_mask_options})
+                    'vector_mask_options': vector_mask_options,
+                    'gdal_warp_options': gdal_warp_options,})
             result_list.append(result)
         worker_pool.close()
         for index, result in enumerate(result_list):
@@ -1714,7 +1715,8 @@ def warp_raster(
         base_raster_path, target_pixel_size, target_raster_path,
         resample_method, target_bb=None, base_sr_wkt=None, target_sr_wkt=None,
         gtiff_creation_options=DEFAULT_GTIFF_CREATION_OPTIONS,
-        n_threads=None, vector_mask_options=None):
+        n_threads=None, vector_mask_options=None,
+        gdal_warp_options=None):
     """Resize/resample raster to desired pixel size, bbox and projection.
 
     Parameters:
@@ -1753,6 +1755,9 @@ def warp_raster(
                     be used to filter the geometry in the mask. Ex:
                     'id > 10' would use all features whose field value of
                     'id' is > 10.
+        gdal_warp_options (list): if present, the contents of this list are
+            passed to the `warpOptions` parameter of `gdal.Warp`. See the
+            GDAL Warp documentation for details.
 
     Returns:
         None
@@ -1819,6 +1824,8 @@ def warp_raster(
     warp_options = []
     if n_threads:
         warp_options.append('NUM_THREADS=%d' % n_threads)
+    if gdal_warp_options:
+        warp_options.extend(gdal_warp_options)
 
     mask_vector_path = None
     mask_layer_name = None